+++ falco-js/src/index.d.ts
@@ -3,8 +3,13 @@ export interface LintOptions {
   autoAddIncludes?: boolean;
   diagnosticsOnly?: boolean;
   deserialize?: boolean;
+  falcoPath?: string;
 }
 
-export function falco(args: string[]): Promise<string>;
+export interface FalcoOptions {
+  falcoPath?: string;
+}
+
+export function falco(args: string[], options?: FalcoOptions): Promise<string>;
 export function lint(file: string): Promise<string>;
 export function lintText(text: string, options?: LintOptions): Promise<unknown>;